refactor: move GamePersistenceService to Sudoku.Core for testability
